feat: remove district page on canada
