+++ src/app/ca/races/state/[stateCode]/page.tsx
@@ -96,7 +96,7 @@ export default async function LocationStateSpecificPage({
       <LocationRaces.KeyRacesStates title={`${stateDisplayName} Constituencies`}>
         {dtsiResults.primaryDistricts.map(district => (
           <NestedPageLink
-            href={urls.locationDistrictSpecific({
+            href={urls.locationDistrictSpecificHouseOfCommons({
               stateCode: validatedStateCode,
               district,
             })}
